Add test.h
