+++ tests/src/test_vector.cpp
@@ -1,22 +1,5 @@
-#include <dbg_macro/dbg.h>
 #include <vector>
-
-template <typename L, typename R>
-void assert_eq(const L& lhs, const R& rhs) {
-  if (lhs != rhs) {
-    std::cerr << "Assertion 'lhs == rhs' failed!\n";
-    dbg(lhs);
-    dbg(rhs);
-    std::exit(1);
-  }
-}
-
-template <typename T>
-std::string pretty_print(T&& value) {
-  std::stringstream stream;
-  dbg_macro::pretty_print(stream, std::forward<T>(value));
-  return stream.str();
-}
+#include "test.h"
 
 int main(int argc, char const* argv[]) {
   const std::vector<int> dummy_vec_int{3, 2, 3};
